+++ packages/env/lib/commands/destroy.js
@@ -17,21 +17,21 @@ const rimraf = util.promisify( require( 'rimraf' ) );
  * Internal dependencies
  */
 const { loadConfig } = require( '../config' );
+const { executeLifecycleScript } = require( '../execute-lifecycle-script' );
 
 /**
  * Destroy the development server.
  *
  * @param {Object}  options
  * @param {Object}  options.spinner A CLI spinner which indicates progress.
+ * @param {boolean} options.scripts Indicates whether or not lifecycle scripts should be executed.
  * @param {boolean} options.debug   True if debug mode is enabled.
  */
-module.exports = async function destroy( { spinner, debug } ) {
-	const { dockerComposeConfigPath, workDirectoryPath } = await loadConfig(
-		path.resolve( '.' )
-	);
+module.exports = async function destroy( { spinner, scripts, debug } ) {
+	const config = await loadConfig( path.resolve( '.' ) );
 
 	try {
-		await fs.readdir( workDirectoryPath );
+		await fs.readdir( config.workDirectoryPath );
 	} catch {
 		spinner.text = 'Could not find any files to remove.';
 		return;
@@ -57,10 +57,14 @@ module.exports = async function destroy( { spinner, debug } ) {
 		return;
 	}
 
+	if ( scripts ) {
+		executeLifecycleScript( 'beforeDestroy', config, spinner );
+	}
+
 	spinner.text = 'Removing docker images, volumes, and networks.';
 
 	await dockerCompose.down( {
-		config: dockerComposeConfigPath,
+		config: config.dockerComposeConfigPath,
 		commandOptions: [ '--volumes', '--remove-orphans', '--rmi', 'all' ],
 		log: debug,
 	} );
@@ -70,7 +74,7 @@ module.exports = async function destroy( { spinner, debug } ) {
 	// by this point, which causes rimraf to fail. We need to wait at least 2.5-5s,
 	// but using 10s in case it's dependant on the machine.
 	await new Promise( ( resolve ) => setTimeout( resolve, 10000 ) );
-	await rimraf( workDirectoryPath );
+	await rimraf( config.workDirectoryPath );
 
 	spinner.text = 'Removed WordPress environment.';
 };