+++ packages/env/lib/config/get-config-from-environment-vars.js
@@ -16,10 +16,11 @@ const { checkPort, checkVersion, checkString } = require( './validate-config' );
  * Environment variable configuration.
  *
  * @typedef WPEnvironmentVariableConfig
- * @property {?number}   port       An override for the development environment's port.
- * @property {?number}   testsPort  An override for the testing environment's port.
- * @property {?WPSource} coreSource An override for all environment's coreSource.
- * @property {?string}   phpVersion An override for all environment's PHP version.
+ * @property {?number}                  port             An override for the development environment's port.
+ * @property {?number}                  testsPort        An override for the testing environment's port.
+ * @property {?WPSource}                coreSource       An override for all environment's coreSource.
+ * @property {?string}                  phpVersion       An override for all environment's PHP version.
+ * @property {?Object.<string, string>} lifecycleScripts An override for various lifecycle scripts.
  */
 
 /**
@@ -33,6 +34,7 @@ module.exports = function getConfigFromEnvironmentVars( cacheDirectoryPath ) {
 	const environmentConfig = {
 		port: getPortFromEnvironmentVariable( 'WP_ENV_PORT' ),
 		testsPort: getPortFromEnvironmentVariable( 'WP_ENV_TESTS_PORT' ),
+		lifecycleScripts: getLifecycleScriptOverrides(),
 	};
 
 	if ( process.env.WP_ENV_CORE ) {
@@ -53,15 +55,6 @@ module.exports = function getConfigFromEnvironmentVars( cacheDirectoryPath ) {
 		environmentConfig.phpVersion = process.env.WP_ENV_PHP_VERSION;
 	}
 
-	if ( process.env.WP_ENV_AFTER_SETUP ) {
-		checkString(
-			'environment variable',
-			'WP_ENV_AFTER_SETUP',
-			process.env.WP_ENV_AFTER_SETUP
-		);
-		environmentConfig.afterSetup = process.env.WP_ENV_AFTER_SETUP;
-	}
-
 	return environmentConfig;
 };
 
@@ -70,7 +63,7 @@ module.exports = function getConfigFromEnvironmentVars( cacheDirectoryPath ) {
  *
  * @param {string} varName The environment variable to check (e.g. WP_ENV_PORT).
  *
- * @return {number} The parsed port number
+ * @return {number} The parsed port number.
  */
 function getPortFromEnvironmentVariable( varName ) {
 	if ( ! process.env[ varName ] ) {
@@ -83,4 +76,34 @@ function getPortFromEnvironmentVariable( varName ) {
 	checkPort( 'environment variable', varName, port );
 
 	return port;
+}
+
+/**
+ * Parses the lifecycle script environment variables.
+ *
+ * @return {Object.<string, string>} The parsed lifecycle scripts.
+ */
+function getLifecycleScriptOverrides() {
+	const lifecycleScripts = {};
+
+	// Find all of the lifecycle script overrides and parse them.
+	for ( const env in process.env ) {
+		const match = env.match( /WP_ENV_LIFECYCLE_SCRIPT_([A-Za-z0-9_]+)/ );
+		if ( ! match ) {
+			continue;
+		}
+
+		const scriptValue = process.env[ env ];
+		checkString( 'environment variable', env, scriptValue );
+
+		// Convert the snake_case environment variables into a camelCase lifecycle script key.
+		const scriptKey = match[ 1 ]
+			.toLowerCase()
+			.replace( /(_[a-z])/g, ( group ) =>
+				group.toUpperCase().replace( '_', '' )
+			);
+		lifecycleScripts[ scriptKey ] = scriptValue;
+	}
+
+	return lifecycleScripts;
 }